+++ src/main/java/Task.java
@@ -1,0 +1,22 @@
+// Task.java
+public class Task {
+    private String taskDescription;
+    private boolean isDone;
+
+    public Task(String taskDescription) {
+        this.taskDescription = taskDescription;
+        this.isDone = false;
+    }
+
+    public void markAsDone() {
+        this.isDone = true;
+    }
+
+    public void markAsUndone() {
+        this.isDone = false;
+    }
+
+    public String toString() {
+        return "|" + (isDone ? " 1 " : " 0 ") + "| " + taskDescription;
+    }
+}
