+++ src/main/java/Todo.java
@@ -1,0 +1,11 @@
+// Todo.java
+public class Todo extends Task {
+    public Todo(String description) {
+        super(description);
+    }
+
+    @Override
+    public String toString() {
+        return "T |" + super.toString();
+    }
+}
